apply confirm to remove post and remove comment

+++ src/main/frontend/src/components/Comment/CommentItem.js
@@ -20,6 +20,8 @@ const CommentItem = (props) => {
   };
 
   const deleteComment = () => {
+    const isConfirmed = window.confirm('정말로 댓글을 삭제하시겠습니까?');
+    if (!isConfirmed) return;
     dispatch(commentActions.deleteComment(id));
   };
   const onCancle = () => {
